Date range error of displaying old simhash file when calculating new simhashes fixed.
Cachefile being deleted after the date range has been passed and the user is trying to calculate unique for all. This is done because during a resend of the request, logs window crashes and stats role shows contents of the old date range cachefile. So the old content is getting deleted all together.

There is a small error of the logs screen getting set to 100% before the simhash count is fully done. This might possibly be due to a resend of the ajax request after a timeout and the percentages of the two requests might keep adding up to get it to 100.

The resend might've been to internet connection drops. Needs more testing. Over all, date range functioning appropriately.